requirements.txt
langchain>=0.0.300
openai>=0.27.0
chromadb>=0.3.32
tqdm
python-dotenv
pdfminer.six
pypdf
sentence-transformers
faiss-cpu
transformers>=4.30.0
accelerate
typing-extensions
uvicorn
fastapi
python-multipart
pytest
streamlit
rapidfuzz
python-multipart